feat(menus): macOS application menu, attachment-download module + Save As
- Extract the ~540-line attachment download pipeline out of menus/index.js
  into app/menus/attachmentDownload.js. menus/index.js is a menu builder, not
  a download engine (addresses the altitude finding from review); the module
  is now independently unit-testable.
- macOS menu fix: the app's menu was applied only via BrowserWindow.setMenu(),
  a no-op on macOS, so Mac users got the generic default Electron menu with
  none of the app's items. Apply via Menu.setApplicationMenu on darwin (shared
  applyApplicationMenu helper used by initialize() and updateMenu()), and add
  the standard Edit/View/Window role menus so Cmd+C/V/A and window shortcuts
  work.
- Add "Save Attachment As…" / "Save File from Clipboard Link As…" context-menu
  items: a native save dialog (in the UI layer) picks the path, passed to the
  download module via a new destinationPath option threaded through all three
  save paths (generic fetch, hidden-window download, text extraction).
- Add tests/unit/attachmentDownload.test.js: 12 tests covering the URL/filename
  helpers and the end-to-end generic fetch→save→clipboard path (incl. traversal
  safety, dedup, destinationPath). Full suite 281/281.
